+++ src/lib/yang/tests/translator_utils_unittests.cc
@@ -113,7 +113,7 @@ TEST(YangReprTest, getTest) {
     EXPECT_NO_THROW_LOG(sess.applyChanges());
 
     xpath = "/keatest-module:main/ui64";
-    uint32_t u64(64);
+    uint64_t u64(64);
     value = to_string(u64);
     EXPECT_NO_THROW_LOG(sess.setItem(xpath, value));
     EXPECT_NO_THROW_LOG(sess.applyChanges());
@@ -137,7 +137,7 @@ TEST(YangReprTest, getTest) {
     EXPECT_NO_THROW_LOG(sess.applyChanges());
 
     xpath = "/keatest-module:main/i64";
-    int32_t s64(64);
+    int64_t s64(64);
     value = to_string(s64);
     EXPECT_NO_THROW_LOG(sess.setItem(xpath, value));
     EXPECT_NO_THROW_LOG(sess.applyChanges());
